% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_converters.R
\name{as_array}
\alias{as_array}
\title{Convert sptable to array-like object}
\usage{
as_array(x, dim_names)
}
\arguments{
\item{x}{\code{sptable}}

\item{dim_names}{A named list of named vectors. See examples.}
}
\description{
Convert sptable to array-like object
}
\examples{
d  <- asia[, 1:3]
# dm <- lapply(d, function(x) structure(unique(x), names = toupper(unique(x))))
# sp <- sptable(d)
# as_array(sp, dm)
}
