% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tables.R
\name{as_sptable}
\alias{as_sptable}
\alias{as_sptable.array}
\alias{as_sptable.matrix}
\alias{as_sptable.table}
\title{Array to Sparse Table}
\usage{
as_sptable(x, validate = TRUE)

\method{as_sptable}{array}(x, validate = TRUE)

\method{as_sptable}{matrix}(x, validate = TRUE)

\method{as_sptable}{table}(x, validate = TRUE)
}
\arguments{
\item{x}{An \code{array}-like object}

\item{validate}{Logical. Check if x has proper dimnames.}
}
\description{
Convert conditional probability table of array-type to a sparse representation
}
\details{
Use only validate = FALSE if you are positve that the structure is correct.
}
\examples{


# Convert one-dimensional array

dmx <- list(X = c("a", "b"))
x   <- array(c(1, 2), dimnames = dmx)
as_sptable(x)

# Convert mutli-dimensional array

dmy <- list(Y = c("a", "b"), Z = c("c", "d"), W = c("e", "f"))
y   <- array(c(1,0,3,0,0,2,1,0), c(2,2,2), dimnames = dmy)
as_sptable(y)

# Convert a data frame to sptable

as_sptable(table(asia[, 2:3]))


}
