% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_env.R
\name{as_sptable_env}
\alias{as_sptable_env}
\alias{as_sptable_env.array}
\alias{as_sptable_env.matrix}
\alias{as_sptable_env.table}
\title{Convert array-like objev to sptable_env}
\usage{
as_sptable_env(x, validate = TRUE)

\method{as_sptable_env}{array}(x, validate = TRUE)

\method{as_sptable_env}{matrix}(x, validate = TRUE)

\method{as_sptable_env}{table}(x, validate = TRUE)
}
\description{
Convert array-like objev to sptable_env
}
