% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asia}
\alias{asia}
\title{Asia}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5000 rows and 8 columns.
}
\usage{
asia
}
\description{
Small synthetic data set from Lauritzen and Spiegelhalter (1988) about lung
diseases (tuberculosis, lung cancer or bronchitis) and visits to Asia. This
copy of the data was taken from the R package "bnlearn" where all values
"yes" have been converted to "y" and all values "no" have been converted
to "n".
}
\details{
\describe{
  \item{\code{D} (dysponea)}{}
  \item{\code{T} (tuberculosis)}{}
  \item{\code{L} (lung cancer)}{}
  \item{\code{B} (bronchitis)}{}
  \item{\code{A} (visit to Asia)}{}
  \item{\code{S} (smoking)}{}
  \item{\code{X} (chest C-ray)}{}
  \item{\code{E} (tuberculosis vs cancer/bronchitis)}{}
}
}
\references{
\href{https://www.bnlearn.com/documentation/man/asia.html}{bnlearn-asia}
}
\keyword{datasets}
