% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asia2}
\alias{asia2}
\title{Asia bnlearn}
\format{
An object of class \code{bn.fit} (inherits from \code{bn.fit.dnet}) of length 8.
}
\usage{
asia2
}
\description{
See the \code{asia} data for information. This version, has class \code{bn.fit}.
}
\references{
\href{https://www.bnlearn.com/bnrepository/discrete-small.html#asia}{bnlearn-asia}
}
\keyword{datasets}
