% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bnlearn.R
\name{bnfit_to_cpts}
\alias{bnfit_to_cpts}
\title{bnfit to cpts}
\usage{
bnfit_to_cpts(x)
}
\arguments{
\item{x}{A \code{bn.fit} object}
}
\description{
Convert a \code{bn.fit} object (a list of cpts from the bnlearn package)
into a list of ordinary array-like cpts
}
