% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_lookup.R
\name{char_array}
\alias{char_array}
\alias{char_array.matrix}
\alias{char_array.table}
\title{Character array}
\usage{
char_array(x)

\method{char_array}{matrix}(x)

\method{char_array}{table}(x)
}
\arguments{
\item{x}{array, matrix or table}
}
\description{
Convert an array-like object into a new representation where all values
are a single character.
}
\details{
\code{sptable} objects only allow for single
character-representations and \code{char_frame} helps achieve this
}
\examples{

ca <- char_array(HairEyeColor)

# retrieve the old dimnames:
lu <- lookup(ca)

# find a particular combination
find(lu, c(Eye = "Brown", Sex = "Male", Hair = "Red"))
}
