% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_lookup.R
\name{char_frame}
\alias{char_frame}
\alias{char_frame.data.frame}
\title{Character frame}
\usage{
char_frame(x)

\method{char_frame}{data.frame}(x)
}
\arguments{
\item{x}{data.frame or matrix}
}
\description{
Convert a data frame or matrix into a new representation where all values
are a single character.
}
\details{
\code{sptable} objects only allow for single
character-representations and \code{char_frame} helps achieve this.
Using \code{lookup} in a \code{char_frame} yields a lookup table
such that new and old values can be compared and looked up. To find
a particular value use \code{find.}
}
\examples{

# Note: asia is already on correct form, so
# no need to use char_frame in principle

ca <- char_frame(asia[, 1:3])
sptable(as.matrix(ca))

# retrieve the old dimnames:
lu <- lookup(ca)

# find a particular combination
find(lu, c(A = "y", S = "y", T = "n"))
}
