% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{compile}
\alias{compile}
\alias{compile.cpt_list}
\title{Compile information}
\usage{
compile(x, root_node = "", save_graph = FALSE)

\method{compile}{cpt_list}(x, root_node = "", save_graph = FALSE)
}
\arguments{
\item{x}{An object returned from \code{cpt_list}}

\item{root_node}{A node for which we require it to live in the root
clique (the first clique)}

\item{save_graph}{Logical.}
}
\description{
Compiled objects are used as building blocks for junction tree inference
}
\examples{
cptl <- cpt_list(asia2)
compile(cptl)
}
