% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{cpt_list}
\alias{cpt_list}
\title{Conditional probability list}
\usage{
cpt_list(x)
}
\arguments{
\item{x}{A named list with cpts in form of array-like object(s).
The name must be the child.}
}
\description{
A check and conversion of cpts to be used in the junction tree algorithm
}
\examples{
print(asia2)
cpt_list(asia2)
}
