% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_lookup.R
\name{find}
\alias{find}
\alias{find.lookup}
\title{Find}
\usage{
find(x, y)

\method{find}{lookup}(x, y)
}
\arguments{
\item{x}{lookup object returned from char_frame or char_array}

\item{y}{named vector}
}
\description{
Find
}
