% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{get_graph}
\alias{get_graph}
\alias{get_graph.charge}
\alias{get_graph.cpt_list}
\title{Get graph}
\usage{
get_graph(x)

\method{get_graph}{charge}(x)

\method{get_graph}{cpt_list}(x)
}
\arguments{
\item{x}{\code{cpt_list} or a compiled object}
}
\value{
A graph as an \code{igraph} object
}
\description{
Retrieve the graph from
}
