% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{get_triang_graph}
\alias{get_triang_graph}
\title{Get triangulated graph}
\usage{
get_triang_graph(x)
}
\arguments{
\item{x}{A compiled object}
}
\value{
A triangulated graph as a neibor matrix
}
\description{
Retrieve the triangulated graph from
}
