% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{dim_names}
\alias{dim_names}
\alias{dim_names.cpt_list}
\alias{names.cpt_list}
\alias{dim_names.charge}
\alias{names.charge}
\title{Cpt list getters}
\usage{
dim_names(x)

\method{dim_names}{cpt_list}(x)

\method{names}{cpt_list}(x)

\method{dim_names}{charge}(x)

\method{names}{charge}(x)
}
\arguments{
\item{x}{\code{cpt_list} or a compiled object}
}
\description{
Getter methods for cpt list objects
}
