% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_jt.R
\name{jt}
\alias{jt}
\title{Junction Tree}
\usage{
jt(g, data, evidence = NULL, flow = "sum", propagate = TRUE, validate = TRUE)
}
\arguments{
\item{g}{Either a decomposable undirected graph as a named list or a \code{gengraph} object or an \code{igraph} DAG}

\item{data}{A \code{data.frame} of characters. Note: all values must only be a single character. See details.}

\item{evidence}{A named vector}

\item{flow}{Character. Either "sum" or "max"}

\item{propagate}{Logical}

\item{validate}{Logical. See details.}
}
\value{
A \code{jt} object
}
\description{
Construction of a junction tree and compilation of the network
}
\details{
It is assumed that all values in \code{data}, for all variables,
are represented as a single character. If \code{validate} is \code{TRUE} this is checked.
If values are not single characters, one may exploit the \code{to_single_chars} function.
}
\examples{

# Setting up the network
# ----------------------

library(igraph)
el <- matrix(c(
"A", "T",
"T", "E",
"S", "L",
"S", "B",
"L", "E",
"E", "X",
"E", "D",
"B", "D"),
 nc = 2,
 byrow = TRUE)

g <- graph_from_edgelist(el)
plot(g)
# -----------------------

# Data
# ----
# We use the asia data; the the man page (?asia)

# Example 1: sum-flow without evidence
# ------------------------------------
jt1 <- jt(g, asia)
plot(jt1)
print(jt1)
query_belief(jt1, c("E", "L", "T"))
query_belief(jt1, c("B", "D", "E"), type = "joint")

# Example 2: sum-flow with evidence
# ---------------------------------
e2  <- c(A = "y", X = "n")
jt2 <- jt(g, asia, e2)
query_belief(jt2, c("E", "L", "T"))
query_belief(jt2, c("B", "D", "E"), type = "joint")

# Notice that, the configuration "yyn" has changed
# dramatically as a consequence of the evidence

# We can get the probability of the evidence:
query_evidence(jt2)

# Example 3: max-flow without evidence
# ------------------------------------
jt3 <- jt(g, asia, flow = "max")
mpe(jt3)

# Example 4: max-flow with evidence
# ---------------------------------
e4  <- c(T = "y", X = "y", D = "y")
jt4 <- jt(g, asia, e4, flow = "max")
mpe(jt4)

# Notice, that the S, B and E changed from "n" to "y"
# as a consequence of the new evidence e4

# Example 5: investigating the clique graph before propagating
# ------------------------------------------------------------
# TBA

}
\seealso{
\code{\link{query_belief}}, \code{\link{mpe}}, \code{\link{get_cliques}}
}
