% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jti_package.R
\docType{package}
\name{jti-package}
\alias{jti}
\alias{jti-package}
\title{jti: Junction Tree Inference}
\description{
Minimal and memory efficient implementation of the junction tree
algorithm using the Lauritzen-Spiegelhalter scheme.
}
\details{
The main functions are \code{cpt_list}, \code{compile},\code{jt}
and \code{query_belief} which together is sufficient to make
inference using the junction tree algorithm.
}
\references{
Local Computations with Probabilities on Graphical Structures
and Their Application to Expert Systems by S. L. Lauritzen
and D. J. Spiegelhalter (1988). Journal of the Royal Statistical
Society: Series B (Methodological) volume 50, issue 2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlindsk/jti}
}

}
\author{
\strong{Maintainer}: Mads Lindskou \email{mads@math.aau.dk}

}
