% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_lookup.R
\name{lookup}
\alias{lookup}
\alias{lookup.char_frame}
\alias{lookup.char_array}
\title{Lookup}
\usage{
lookup(x)

\method{lookup}{char_frame}(x)

\method{lookup}{char_array}(x)
}
\arguments{
\item{x}{lookup object returned from char_frame or char_array}
}
\description{
Lookup
}
