% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_sptable_unity.R
\name{make_unity_sptable}
\alias{make_unity_sptable}
\title{Sparse unity table}
\usage{
make_unity_sptable(x)
}
\arguments{
\item{x}{Named list with the levels of the corresponding variables}
}
\description{
Sparse unity table
}
