% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_env.R
\name{marginalize}
\alias{marginalize}
\alias{marginalize.sptable_env}
\title{Marginalize}
\usage{
marginalize(p, s, flow = "sum")

\method{marginalize}{sptable_env}(p, s, flow = "sum")
}
\arguments{
\item{p}{A \code{sptable} object}

\item{s}{The variables to be marginalized out}

\item{flow}{Either "sum" or "max" where "max"}
}
\description{
Marginalize
}
\examples{
p <- sptable_env(as.matrix(asia[, 3:5]))
p
marginalize(p, "L")
}
