% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_merge_and_marg.R
\name{\%m*\%}
\alias{\%m*\%}
\alias{\%m/\%}
\alias{\%m+\%}
\alias{\%m-\%}
\title{Sparse table algebra}
\usage{
x \%m*\% y

x \%m/\% y
}
\arguments{
\item{x, y}{\code{sptable} objects}
}
\description{
Multiplication, division, etc. for sparse tables
}
