% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_merge_and_marg.R
\name{merge.sptable}
\alias{merge.sptable}
\title{Merge sparse tables}
\usage{
\method{merge}{sptable}(x, y, op = "*", ...)
}
\arguments{
\item{x}{A \code{sptable} object}

\item{y}{A \code{sptable} object}

\item{op}{Either "*" (multiplication) or "/" (division)}

\item{...}{For S3 compatability}
}
\description{
Multiplication, division, addition and subtraction of two sparse tables
}
\examples{
# Variables in common
x1 <- sptable(asia[, 2:5])
y1 <- sptable(asia[, c(5, 8, 4)])
merge(x1, y1, "*")

# No variables in common
x2 <- sptable(asia[, 1:2])
y2 <- sptable(asia[, 3, drop = FALSE])
x2 \%m/\% y2

}
\seealso{
\code{\link{\%m*\%}}, \code{\link{\%m/\%}}
}
