% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_jt.R
\name{mpe}
\alias{mpe}
\alias{mpe.jt}
\title{Most Probable Explanation}
\usage{
mpe(x)

\method{mpe}{jt}(x)
}
\arguments{
\item{x}{A junction tree object, \code{jt}, with max-flow.}
}
\description{
Returns the most probable explanation given the evidence entered in the
junction tree
}
\examples{
# See the 'jt' function
}
\seealso{
\code{\link{jt}}
}
