% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{munin}
\alias{munin}
\title{Munin}
\format{
An object of class \code{list} of length 1041.
}
\usage{
munin
}
\description{
A massive bayesian network
}
\references{
\href{https://www.bnlearn.com/bnrepository/discrete-massive.html#munin}{bnlearn-munin}
}
\keyword{datasets}
