% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_jt.R
\name{plot.charge}
\alias{plot.charge}
\title{A plot method for junction trees}
\usage{
\method{plot}{charge}(x, ...)
}
\arguments{
\item{x}{A compile object}

\item{...}{For S3 compatability. Not used.}
}
\description{
A plot method for junction trees
}
\seealso{
\code{\link{compile}}
}
