% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_jt.R
\name{plot.jt}
\alias{plot.jt}
\title{A plot method for junction trees}
\usage{
\method{plot}{jt}(x, ...)
}
\arguments{
\item{x}{A junction tree object, \code{jt}.}

\item{...}{For S3 compatability. Not used.}
}
\description{
A plot method for junction trees
}
\seealso{
\code{\link{jt}}
}
