% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{print.charge}
\alias{print.charge}
\title{A print method for compiled objects}
\usage{
\method{print}{charge}(x, ...)
}
\arguments{
\item{x}{A compiled object}

\item{...}{For S3 compatability. Not used.}
}
\description{
A print method for compiled objects
}
\seealso{
\code{\link{jt}}
}
