% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{print.cpt_list}
\alias{print.cpt_list}
\title{A print method for cpt lists}
\usage{
\method{print}{cpt_list}(x, ...)
}
\arguments{
\item{x}{A \code{cpt_list} object}

\item{...}{For S3 compatability. Not used.}
}
\description{
A print method for cpt lists
}
\seealso{
\code{\link{compile}}
}
