% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_jt.R
\name{print.jt}
\alias{print.jt}
\title{A print method for junction trees}
\usage{
\method{print}{jt}(x, ...)
}
\arguments{
\item{x}{A junction tree object, \code{jt}.}

\item{...}{For S3 compatability. Not used.}
}
\description{
A print method for junction trees
}
\seealso{
\code{\link{jt}}
}
