% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{print.pot_list}
\alias{print.pot_list}
\title{A print method for pot lists}
\usage{
\method{print}{pot_list}(x, ...)
}
\arguments{
\item{x}{A \code{pot_list} object}

\item{...}{For S3 compatability. Not used.}
}
\description{
A print method for pot lists
}
\seealso{
\code{\link{compile}}
}
