% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_misc_methods.R
\name{print.sptable}
\alias{print.sptable}
\title{Print}
\usage{
\method{print}{sptable}(x, ...)
}
\arguments{
\item{x}{sptable}

\item{...}{For S3 compatability.}
}
\description{
Print
}
\examples{
sp <- sptable(asia[, 1:3])
print(sp)
}
