% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_jt.R
\name{propagate}
\alias{propagate}
\alias{propagate.jt}
\title{Propagation of junction trees}
\usage{
propagate(x, prop = "full")

\method{propagate}{jt}(x, prop = "full")
}
\arguments{
\item{x}{A junction tree object \code{jt}}

\item{prop}{Either "collect" or "full".}
}
\description{
Given a junction tree object, propagation is conducted
}
\examples{
# See Example 1 in the 'jt' function
}
\seealso{
\code{\link{jt}}
}
