% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_jt.R
\name{query_evidence}
\alias{query_evidence}
\alias{query_evidence.jt}
\title{Query Evidence}
\usage{
query_evidence(x)

\method{query_evidence}{jt}(x)
}
\arguments{
\item{x}{A junction tree object, \code{jt}.}
}
\description{
Get the probability of the evidence entered in the junction tree object
}
\examples{
# See the 'jt' function
}
\seealso{
\code{\link{jt}}, \code{\link{mpe}}
}
