% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_jt.R
\name{send_messages}
\alias{send_messages}
\title{Send Messages in a Junction Tree}
\usage{
send_messages(jt)
}
\arguments{
\item{jt}{A \code{jt} object return from the \code{jt} function}
}
\description{
Send messages from the current leaves to the current parents
in a junction tree
}
\examples{
# See example 6 in the help page for the jt function
}
\seealso{
\code{\link{jt}}, \code{\link{get_cliques}}, \code{\link{jt_leaves}},
\code{\link{jt_parents}}
}
