% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_jt.R
\name{set_evidence}
\alias{set_evidence}
\alias{set_evidence.jt}
\title{Enter Evidence}
\usage{
set_evidence(x, evidence)

\method{set_evidence}{jt}(x, evidence)
}
\arguments{
\item{x}{A junction tree object, \code{jt}.}

\item{evidence}{A named vector. The names are the variabes and the elements
are the evidence.}
}
\description{
Enter evidence into a the junction tree object that has not been propagated
}
\examples{
# See the 'jt' function
}
\seealso{
\code{\link{jt}}, \code{\link{mpe}}
}
