% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_sim.R
\name{sim_data_from_dmrf}
\alias{sim_data_from_dmrf}
\title{Simulate data from a decomposable discrete markov random field}
\usage{
sim_data_from_dmrf(
  graph,
  lvls,
  nsims = 1000,
  increasing_prob = FALSE,
  p1 = 0.8,
  p2 = 1
)
}
\arguments{
\item{graph}{A decomposable discrete markov random field as an igraph object}

\item{lvls}{Named integer vector where each element is the size of the
statespace of the corresponding variable}

\item{nsims}{Number of simulations
distributions from which the simulatios are drawn.}

\item{increasing_prob}{Logical. If true, probabilities in the underlying CPTs
increases with as the number of levels increses.}

\item{p1}{Probability}

\item{p2}{Probability}
}
\description{
Simulate data from a decomposable discrete markov random field
}
