% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_env.R
\name{sptable_env}
\alias{sptable_env}
\title{Sparse table}
\usage{
sptable_env(x)
}
\arguments{
\item{x}{Character matrix}
}
\description{
A sparse contingency table representation of a matrix.
}
\examples{
sptable(as.matrix(asia[, 1:3]))
}
\seealso{
\code{\link{as_parray}}
}
