% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_lookup.R
\name{[.char_frame}
\alias{[.char_frame}
\title{Subset char frame}
\usage{
\method{[}{char_frame}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{\code{char_frame} object}

\item{i}{First index for subsetting. Behaves as an ordinary data frame}

\item{j}{Second index for subsetting. Behaves as an ordinary data frame}

\item{...}{For S3 compatability}

\item{drop}{If TRUE the result is coerced to the lowest possible dimension}
}
\description{
Subset char frame
}
