% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_misc_methods.R
\name{[.sptable}
\alias{[.sptable}
\title{Subset sptable}
\usage{
\method{[}{sptable}(x, i, result = "sptable")
}
\arguments{
\item{x}{sptable}

\item{i}{character vector}

\item{result}{'sptable' or 'value'. The former returns a new
sptable and the latter returns a vector of counts.}
}
\description{
Subset sptable
}
\examples{
sp <- sptable(as.matrix(asia[, 1:3]))
sp[c("yny", "nyy")]
sp[c("yny", "nyy"), "value"]
}
