% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable.R
\name{to_cpt}
\alias{to_cpt}
\alias{to_cpt.sptable}
\title{Conditional probability table}
\usage{
to_cpt(x, y = NULL)

\method{to_cpt}{sptable}(x, y = NULL)
}
\arguments{
\item{x}{A sparse table obtained from \code{sptable}}

\item{y}{Character vector with variables to condition on}
}
\value{
A \code{sptable} object
}
\description{
Returns a sparse conditional probability table conditioned on variables in \code{b}.
}
\details{
If \code{y} is \code{NULL}, \code{x} is just converted to a \code{sptable} with no conditioning variables, i.e. the marginal table.
}
\examples{
sp  <- sptable(as.matrix(asia[, 1:3]))
psp <- to_cpt(sp, c("S", "T"))
sum(psp) # Since (S,T) have 4 configurations
}
\seealso{
\code{\link{sptable}}
}
