% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_misc_methods.R
\name{sum.sptable}
\alias{sum.sptable}
\alias{max.sptable}
\alias{min.sptable}
\alias{which_min}
\alias{which_min.sptable}
\alias{which_max}
\alias{which_max.sptable}
\title{Vector-like operations on sptables}
\usage{
\method{sum}{sptable}(x, ...)

\method{max}{sptable}(x, ...)

\method{min}{sptable}(x, ...)

which_min(x)

\method{which_min}{sptable}(x)

which_max(x)

\method{which_max}{sptable}(x)
}
\arguments{
\item{x}{sptable}

\item{...}{For S3 compatability.}
}
\description{
Vector-like operations on sptables
}
\examples{
sp <- sptable(as.matrix(asia[, 1:3]))
print(sp)
sum(sp) # = nrow(asia)
min(sp)
max(sp)
which_min(sp)
which_max(sp)
}
