% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable.R
\name{as_sptable}
\alias{as_sptable}
\alias{as_sptable.array}
\alias{as_sptable.matrix}
\alias{as_sptable.table}
\title{Convert array-like objev to sptable}
\usage{
as_sptable(x, validate = TRUE)

\method{as_sptable}{array}(x, validate = TRUE)

\method{as_sptable}{matrix}(x, validate = TRUE)

\method{as_sptable}{table}(x, validate = TRUE)
}
\description{
Convert array-like objev to sptable
}
