% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_converters.R
\name{as_sptable}
\alias{as_sptable}
\alias{as_sptable.array}
\alias{as_sptable.matrix}
\alias{as_sptable.table}
\title{Convert an array-like object to sptable}
\usage{
as_sptable(x, validate = TRUE)

\method{as_sptable}{array}(x, validate = TRUE)

\method{as_sptable}{matrix}(x, validate = TRUE)

\method{as_sptable}{table}(x, validate = TRUE)
}
\arguments{
\item{x}{array-like type}

\item{validate}{Boolean. Check if \code{x} has valid dimnames}
}
\description{
Convert an array-like object to sptable
}
\examples{
ca <- char_array(asia2[[4]])
sp <- as_sptable(ca)
sp

# char_array is used to turn the array into correct form, i.e.
# correct length of dimnames

# retrieve the old dimnames:
lu <- lookup(ca)

# find a particular combination
x <- find(lu, c(lung = "yes", smoke = "no"))
sp[paste(unlist(x), collapse = "")]          # Preserving sptable class
sp[paste(unlist(x), collapse = ""), "value"] # returning the value
}
