% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{compile}
\alias{compile}
\alias{compile.cpt_list}
\alias{compile.data.frame}
\title{Compile information}
\usage{
compile(x, g = NULL, root_node = "")

\method{compile}{cpt_list}(x, g = NULL, root_node = "")

\method{compile}{data.frame}(x, g, root_node = "")
}
\arguments{
\item{x}{Either a \code{data.frame} or an object returned from \code{cpt_list}}

\item{g}{Either a directed acyclic graph (DAG) as an igraph object or a
decomposable graph obtained from \code{ess::fit_graph} or an
adjacency list - a named list where an element is a character vector defining
the neigbors of the element. If \code{x} is a \code{cpt_list},
\code{g} must be \code{NULL}. The procdure then deduce the graph
from the conditional probability tables that was inputtet in \code{cpt_list}}

\item{root_node}{A node for which we require it to live in the root clique (the first clique).}
}
\description{
Compiled objects are used as building blocks for junction tree inference
}
