% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{dim_names}
\alias{dim_names}
\alias{dim_names.cpt_list}
\alias{names.cpt_list}
\alias{dim_names.pot_list}
\alias{names.pot_list}
\alias{dim_names.charge}
\alias{names.charge}
\alias{dim_names.jt}
\alias{names.jt}
\title{Variable getters}
\usage{
dim_names(x)

\method{dim_names}{cpt_list}(x)

\method{names}{cpt_list}(x)

\method{dim_names}{pot_list}(x)

\method{names}{pot_list}(x)

\method{dim_names}{charge}(x)

\method{names}{charge}(x)

\method{dim_names}{jt}(x)

\method{names}{jt}(x)
}
\arguments{
\item{x}{\code{cpt_list} or a compiled object}
}
\description{
Getter methods for information about variables and their statespaces
}
