% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_compile.R
\name{dim_names}
\alias{dim_names}
\alias{has_inconsistencies}
\alias{dim_names.cpt_list}
\alias{names.cpt_list}
\alias{dim_names.charge}
\alias{names.charge}
\alias{has_inconsistencies.charge}
\alias{dim_names.jt}
\alias{names.jt}
\alias{has_inconsistencies.jt}
\title{Various getters}
\usage{
dim_names(x)

has_inconsistencies(x)

\method{dim_names}{cpt_list}(x)

\method{names}{cpt_list}(x)

\method{dim_names}{charge}(x)

\method{names}{charge}(x)

\method{has_inconsistencies}{charge}(x)

\method{dim_names}{jt}(x)

\method{names}{jt}(x)

\method{has_inconsistencies}{jt}(x)
}
\arguments{
\item{x}{\code{cpt_list}, \code{pot_list}, \code{charge} or \code{jt}}
}
\description{
Getter methods for \code{cpt_list}, \code{pot_list}, \code{charge}
and \code{jt} objects
}
