% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_set_evidence.R
\name{initialize}
\alias{initialize}
\title{Initialize}
\usage{
initialize(x)
}
\arguments{
\item{x}{A compiled object.}
}
\description{
Initialization of CPTs
}
\details{
Multiply the CPTs and allocate them to clique potentials.
}
