% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_binary_ops.R
\name{jt_nbinary_ops}
\alias{jt_nbinary_ops}
\alias{jt_nbinary_ops.jt}
\alias{jt_nbinary_ops.triangulation}
\title{Number of Binary Operations}
\usage{
jt_nbinary_ops(x, evidence = character(0), root = NULL)

\method{jt_nbinary_ops}{jt}(x, evidence = character(0), root = NULL)

\method{jt_nbinary_ops}{triangulation}(x, evidence = character(0), root = NULL)
}
\arguments{
\item{x}{A junction tree object or an object returned from
the triangulation function}

\item{evidence}{Character vector of evidence nodes}

\item{root}{Integer specifying the root node in the junction tree}
}
\description{
Number of binary operations needed to propagate in a junction tree
given evidence, using the Lauritzen-Spiegelhalter scheme
}
