% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jti_package.R
\docType{package}
\name{jti-package}
\alias{jti}
\alias{jti-package}
\title{jti: Junction Tree Inference}
\description{
Minimal and memory efficient implementation of the junction tree
algorithm using the Lauritzen-Spiegelhalter scheme. S. L. Lauritzen
and D. J. Spiegelhalter (1988).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlindsk/jti}
}

}
\author{
\strong{Maintainer}: Mads Lindskou \email{mads@math.aau.dk}

}
