% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jti.R
\docType{package}
\name{jti-package}
\alias{jti}
\alias{jti-package}
\title{jti: Junction Tree Inference}
\description{
This package is a fast implementation of the junction tree algorithm using the LS-updating scheme.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlindsk/jti}
}

}
\author{
\strong{Maintainer}: Mads Lindskou \email{mads@math.aau.dk}

}
