% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable_merge_and_marg.R
\name{marginalize}
\alias{marginalize}
\alias{marginalize.sptable}
\title{Marginalize}
\usage{
marginalize(p, s, flow = "sum")

\method{marginalize}{sptable}(p, s, flow = "sum")
}
\arguments{
\item{p}{A \code{sptable} object}

\item{s}{The variables to be marginalized out}

\item{flow}{Either "sum" or "max"}
}
\description{
Marginalize
}
\examples{
p <- sptable(asia[, 3:5])
p
marginalize(p, "L")
}
