% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tables.R
\name{merge.sptable}
\alias{merge.sptable}
\title{Merge sparse tables}
\usage{
\method{merge}{sptable}(x, y, op = "*", ...)
}
\arguments{
\item{x}{A \code{sptable} object}

\item{y}{A \code{sptable} object}

\item{op}{Either "*" (multiplication) or "/" (division)}

\item{...}{Not used. For S3 compatability}
}
\description{
Multiplication or division of two sparse tables
}
\examples{
x <- sptable(as.matrix(asia[, 3:5]))
y <- sptable(as.matrix(asia[, 4:6]))
x
y
merge(x, y, "/")
}
