% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable.R
\name{merge.sptable}
\alias{merge.sptable}
\title{Merge sparse tables}
\usage{
\method{merge}{sptable}(x, y, op = "*", validate = TRUE, ...)
}
\arguments{
\item{x}{A \code{sptable} object}

\item{y}{A \code{sptable} object}

\item{op}{Either "*" (multiplication), "/" (division), "+" (addition) or "-" (subtraction)}

\item{...}{Not used. For S3 compatability}
}
\description{
Multiplication or division of two sparse tables
}
\examples{

# Variables in common
# -------------------
x1 <- sptable(as.matrix(asia[, 2:5]))
y1 <- sptable(as.matrix(asia[, c(5, 8, 4)]))
merge(x1, y1, "/")

# No variables in common
x2 <- sptable(as.matrix(asia[, 1:2]))
y2 <- sptable(as.matrix(asia[, 3, drop = FALSE]))
merge(x2, y2, "/")

# Names not on correct form
z  <- chickwts[, 2, drop = FALSE]
x3 <- sptable(as.matrix(z))
merge(x3, x3)

# Correcting the names
z_new <- to_chars(z)
x4    <- sptable(as.matrix(z_new))
x4

}
