% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_set_evidence.R
\name{set_evidence}
\alias{set_evidence}
\alias{set_evidence.jt}
\alias{set_evidence.charge}
\title{Enter Evidence}
\usage{
set_evidence(x, evidence, initialize_cpts = FALSE)

\method{set_evidence}{jt}(x, evidence, initialize_cpts = FALSE)

\method{set_evidence}{charge}(x, evidence, initialize_cpts = TRUE)
}
\arguments{
\item{x}{A junction tree object, \code{jt}.}

\item{evidence}{A named vector. The names are the variabes and the elements
are the evidence.}

\item{initialize_cpts}{\code{TRUE} if the CPTs should be initialized and then
create the clique potentials. Only relevant on objects returned from \code{compile}.}
}
\description{
Enter evidence into a the junction tree object that has not been propagated
}
\examples{
# See the 'jt' function
}
\seealso{
\code{\link{jt}}, \code{\link{mpe}}
}
