% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sptable.R
\name{sptable}
\alias{sptable}
\title{Sparse table}
\usage{
sptable(x, validate = TRUE)
}
\arguments{
\item{x}{Character matrix}

\item{validate}{Logical. If TRUE, it checks whether or not the values of all variables
in \code{A} are constrained to a single character. If not, an error is produced.}
}
\description{
A sparse contingency table representation of a matrix.
}
\details{
The reason for the values to be constrained to a single character is due to
an increase in performance.
}
\examples{
sptable(as.matrix(asia[, 1:3]))
}
\seealso{
\code{\link{to_cpt}}
}
