% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_triang.R
\name{triangulate}
\alias{triangulate}
\alias{triangulate.cpt_list}
\title{Triangulate a Bayesian network}
\usage{
triangulate(
  x,
  joint_vars = NULL,
  tri = "min_fill",
  evidence_nodes = character(0)
)

\method{triangulate}{cpt_list}(
  x,
  joint_vars = NULL,
  tri = "min_fill",
  evidence_nodes = character(0)
)
}
\arguments{
\item{x}{An object returned from \code{cpt_list}}

\item{joint_vars}{A vector of variables for which we require them
to be in the same clique. Edges between all these variables are added
to the moralized graph.}

\item{tri}{The optimization strategy used for triangulation. Either
one of 'min_nei', 'min_fill', 'min_sp', 'evidence', 'minimal'}

\item{evidence_nodes}{Character vector. TODO: More details}
}
\description{
Given a list of CPTs, this function finds a triangulation
}
